backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart

langchain
langchain-community
langchain-text-splitters

chromadb
sentence-transformers
huggingface-hub

groq
pypdf
